requirements.txt
beautifulsoup4==4.12.3
feedparser==6.0.11
Flask==3.0.3
google-generativeai
gTTS==2.5.1
python-dotenv==1.0.1
Werkzeug==3.0.3